+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+- Add support for posting a comment to ClickUp using the `app-version:notify-clickup` command.
+
 ## 3.0.0 (2025-03-25)
 - Updated the Newrelic API to use GraphQL
   - Removed env variable `APP_VERSION_NEWRELIC_APPLICATION_ID`.